ifplugd-conf: add ifplugd configuration files
This will provide the possibility to use ifplugd and ifup/ifdown as
network managment.

The assumption is made that we have up to 3 network interfaces, which is
currently the maximum number of interfaces we have on some MX-4 devices.

This recipe only provides the configuration files for ifplugd. The
seperation is to be able to provide different types of network
managment.

Signed-off-by: Mirza Krak <mirza.krak@endian.se>