Add recipe for development console image
This recipe include gdb, gdbserver,gcc,make and libs for debug.
Maybe other tools will be added later on. But for now it is enough to compile and
debug direct on the machine and using for example Eclipse to debug a new application
direct on the machine (using ssh and populated_sdk to build the "new" app).